+++ docs/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - unified logging across the whole OSIDB
+- validate hightouch and hightouch-lite flag value combinations (OSIDB-329)
 
 ## [2.1.0] - 2022-08-02
 ### Changed