+++ docs/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Treat collector failures due to already running collectors or due to
   waiting for dependencies as celery Retry exceptions.
 - OSIDB now uses publicly available images from docker.io (OSIDB-170)
+- fix bug that Major Incident can be unset by unrelated BZ flag (OSIDB-416)
 
 ### Added
 - support for CVE-less flaws (OSIDB-25)