+++ docs/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [2.2.0] - 2022-09-05
 ### Changed
 - Authentication is no longer compulsory for read-only requests against the
   main OSIDB endpoints such as /flaws, /affects and /trackers (OSIDB-313)